+++ Sources/PDKit/PDConstants.swift
@@ -50,6 +50,7 @@ public class DefaultSettings {
     public static let MentionedDisclaimerRawValue = false
     public static let SiteIndexRawValue = 0
     public static let MaxSupportedNotificationsMinutesBefore = 120
+    public static let PillsEnabledRawValue = true
 }
 
 /// Default properties to assign to pills when they are not given any.